+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Typed QRCode views
 - Typed rate info views
 - Typed rate views
+- Typed review settings views
 
 ### Изменено